duplication in mailsending reduced
